Move JNI functions out of jnifunc.c
The functions can be in their respective modules.
